Add #section tags to UI
Work in progress, this should allow links within the single-page UI

It's worth deciding whether this even needs to be single-page: the
file listings could easily be their own separate page and Tensorboard
already is.

But I like being able to flip between the visuals and the console
log output without re-loading the console.
Maybe the two could be munged together somehow.