Add undocumented gnomehat-run options for demos
Originally, the /demos/ webapi just shelled out to "gnomehat demo",
which would clone a git repository from github.com/gnomehat and then
run a "gnomehat run" job to turn that repository into an experiment.

But some repositories actually take a long time just to clone- either
because of their size or bad network conditions. So we can't just
freeze the UI until the background `git clone` process finishes.
And we can't redirect the user immediately to the main page because
they'll just be confused- it might take 10 or 20 seconds for their
demo to clone and start running.

So basically, the clone job is a long-running background process.
How do we run a background process? Idea: use gnomehat!

So, now the /demos API runs `gnomehat run gnomehat demo`

The `run` just downloads the git repo and the `demo` actually runs it.