Support commands run from subdirectories
Many users have a repository with one or more subdirectories, and their
workflow requires cd'ing into a subdirectory before running a command
(instead of eg. using python -m packagename.module or hacking sys.path)

A great example of this is the Pytorch examples repository, which
obviously should work out of the box with Gnomehat.